+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
-- Dropping support for Ruby < 3.1
+- Dropping support for Ruby < 3.1 [#631](https://github.com/glebm/i18n-tasks/pull/631)
+- OpenAI translator now uses language names instead of locale codes. [#630](https://github.com/glebm/i18n-tasks/pull/630)
 
 ## v1.0.15